requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
drf-nested-routers==0.91
filelock==3.0.12
firebase-admin==4.3.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
identify==1.4.20
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
msgpack==1.0.0
nodeenv==1.4.0
pre-commit==2.5.1
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rope==0.17.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
wrapt==1.12.1
